Making modules folder for gpt
