+++ examples/llm/__init__.py
@@ -7,11 +7,11 @@ try:
     from examples.llm.src.flash_attention import FlashAttention, FlashMHA
     from examples.llm.src.hf_causal_lm import ComposerHFCausalLM
     from examples.llm.src.model_registry import COMPOSER_MODEL_REGISTRY
-    from examples.llm.src.mosaic_gpt import (GPTMLP, ComposerMosaicGPT,
-                                             FlashCausalAttention, GPTBlock,
-                                             MosaicGPT, TorchCausalAttention,
-                                             TritonFlashCausalAttention,
-                                             alibi_bias)
+    from examples.llm.src.modules.attention_modules import (
+        FlashCausalAttention, TorchCausalAttention, TritonFlashCausalAttention,
+        alibi_bias)
+    from examples.llm.src.modules.gpt_blocks import GPTMLP, GPTBlock
+    from examples.llm.src.mosaic_gpt import ComposerMosaicGPT, MosaicGPT
     from examples.llm.src.tokenizer import (TOKENIZER_REGISTRY, HFTokenizer,
                                             LLMTokenizer)
 except ImportError as e:
